# backend/public/api-docs/openapi.yml
openapi: 3.0.3
info:
  title: Ma Commune API
  description: |- 
    API pour l'application mobile et le back-office de Ma Commune.
    Cette specification couvre le socle V1 verifie localement :
    signalements, suivi citoyen, notifications, consultations et agenda communal.
  version: 1.6.0
servers:
  - url: /api
    description: Instance locale ou reverse proxy communal
security:
  - bearerAuth: []
paths:
...
  /incidents/{id}/photos:
    get:
      summary: Lister les photos d'un incident
      tags: [Photos]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Liste des photos
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Photo' }
    post:
      summary: Uploader une photo (max 5)
      tags: [Photos]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo: { type: string, format: binary }
                sort_order: { type: integer }
      responses:
        '201':
          description: Photo uploadée
  /incidents/{id}/photos/{photoId}:
    delete:
      summary: Supprimer une photo
      tags: [Photos]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: photoId, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 'Photo supprimée' }

  # ── Commentaires (v1.4) ───────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Photo:
      type: object
      properties:
        id: { type: integer }
        url: { type: string, format: uri }
        file_name: { type: string }
        sort_order: { type: integer }
...